% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGAS.R
\docType{data}
\name{epa}
\alias{epa}
\title{EPAGAS data set}
\format{

}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William-Mendenhall/dp/1498728855>
}
\usage{
epa
}
\description{
A complete data set of MPG's for some unknown vehicles
}
\keyword{datasets}
