% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybiplot.R
\name{mybiplot}
\alias{mybiplot}
\title{mybiplot}
\usage{
mybiplot(x = double(), k = double())
}
\arguments{
\item{x}{a quantitative vector of continuous values}

\item{k}{a double constant}
}
\value{
an S3 object of type list and class math4753biplotKDeason
}
\description{
mybiplot
}
\examples{
\dontrun{ mibiplot() }
}
