% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.math4753biplotKDeason.R
\name{print.math4753biplotKDeason}
\alias{print.math4753biplotKDeason}
\title{print.math4753biplotKDeason}
\usage{
\method{print}{math4753biplotKDeason}(x, ...)
}
\arguments{
\item{x}{a quantitative vector of continuous values}

\item{...}{additional arguments}
}
\value{
a histogram
}
\description{
print.math4753biplotKDeason
}
\examples{
 \dontrun{print(mybiplot(x = mpg, k = 1))}
}
