% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.math4753biplotKDeason.R
\name{summary.math4753biplotKDeason}
\alias{summary.math4753biplotKDeason}
\title{summary.math4753biplotKDeason}
\usage{
\method{summary}{math4753biplotKDeason}(x, ...)
}
\arguments{
\item{x}{a quantitative vector of continuous values}

\item{...}{additional arguments}
}
\value{
a tibble summarizing x grouping by y
}
\description{
summary.math4753biplotKDeason
}
\examples{
\dontrun{summary(mybiplot(x = mpg, k = 1))}
}
